requirements.txt
asgiref
Django
psycopg2-binary
sqlparse
tzdata
whitenoise
environs
djangorestframework
Pillow
djangorestframework-simplejwt
djoser
drf-spectacular
ruff